Rename Xgit.Plumbing.UpdateIndex.CacheInfo.run/2 to Xgit.Repository.Plumbing.update_index_cache_info/2.
